mcp_server/requirements.txt
# Core API Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Authentication & Security
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Database & Storage
supabase==2.4.4
postgrest==0.16.8

# Elasticsearch for Vector Search
elasticsearch==8.12.0

# AI & Embeddings
google-generativeai==0.5.4

# Semantic Search (Production-Ready Versions)
sentence-transformers==2.7.0
torch==2.2.0
transformers==4.40.2
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2
huggingface-hub==0.23.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
lxml==5.2.1
beautifulsoup4==4.12.3

# MIME Type Detection (Optional - may need system dependencies)
# python-magic==0.4.27  # Uncomment if needed, requires libmagic

